requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
pandas>=2.0.0
pdfplumber==0.10.2
openpyxl==3.1.2
Werkzeug==2.3.7
gunicorn==21.2.0
itsdangerous==2.1.2